refactor: unexport private members in the `QueryClient`
